simplify handshake, fix handling of some messages
